backend/packages/aria-contracts/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aria-contracts"
version = "0.1.0"
description = "Shared Pydantic schemas for ARIA microservices — the single source of truth for inter-service data."
requires-python = ">=3.11"
authors = [{name = "ARIA Team"}]
license = {text = "Apache-2.0"}
dependencies = [
    "pydantic>=2.7,<3",
    "openenv-core>=0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8", "mypy>=1.10"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aria_contracts = ["py.typed"]
